+++ src/main/java/roomescape/global/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package roomescape.global;
+package roomescape.global.exception;
 
 public record ExceptionResponse(
         String message
